added the aggregation into a sub folder for the project Updated the README
