+++ dimagi_lake/aggregation/nutrition_project/aggregation_helpers/child_care_monthly.py
@@ -1,5 +1,5 @@
 from consts import FLWC_LOCATION_TABLE, SERVICE_ENROLLMENT_TABLE, CHILD_WEIGHT_HEIGHT_FORM_TABLE, SUPPLEMENTARY_NUTRITION_FORM_TABLE, CHILD_THR_FORM_TABLE
-from dimagi_lake.aggregation.aggregation_helpers.base_helper import \
+from dimagi_lake.aggregation.nutrition_project.aggregation_helpers.base_helper import \
     BaseAggregationHelper
 from spark_session_handler import SPARK
