+++ dimagi_lake/aggregation/nutrition_project/aggregation_helpers/child_snd_form.py
@@ -1,30 +1,29 @@
-from dimagi_lake.aggregation.aggregation_helpers.base_helper import BaseAggregationHelper
+from dimagi_lake.aggregation.nutrition_project.aggregation_helpers.base_helper import BaseAggregationHelper
 from spark_session_handler import SPARK
 
 
-class ChildTHRAggregationHelper(BaseAggregationHelper):
+class SupplementaryNutritionAggregationHelper(BaseAggregationHelper):
 
     @property
     def source_tablename(self):
-        return f"{self.database_name}.raw_child_thr_form"
+        return f"{self.database_name}.raw_child_snd_form"
 
     @property
     def columns(self):
         return {
             'children_cases': (
                 'form',
-                'children_0_3_cases',
+                'children_cases',
                 'item'
             ),
-            'children_received_thr_id': (
+            'present_children': (
                 'form',
-                'thr_details_group',
-                'children_received_thr_id'
+                'children_attendance',
+                'present_children'
             ),
-            'number_of_days_thr_given': (
+            "children_food_served": (
                 'form',
-                'thr_details_group',
-                'number_of_days_thr_given'
+                'children_food_served'
             ),
             "timeend": (
                 'form',
@@ -40,18 +39,23 @@ class ChildTHRAggregationHelper(BaseAggregationHelper):
         return self.enforce_schema(df)
 
     def preprocess(self):
-        return SPARK.sql(f"""
+        df1 = SPARK.sql(f"""
         SELECT  
             children.`@id` as child_case_id,
             date_trunc('DAY', {self.get_column('timeend')}) as timeend,
-            CASE
-                WHEN array_contains(split({self.get_column('children_received_thr_id')}, ' '), children.`@id`) THEN {self.get_column('number_of_days_thr_given')}
+            CASE 
+                WHEN array_contains(split({self.get_column('present_children')}, ' '), children.`@id`) THEN 1 
+                ELSE 0 
+            END as pse_attended,
+            CASE 
+                WHEN array_contains(split({self.get_column('children_food_served')}, ' '), children.`@id`) THEN 1
                 ELSE 0
-            END as days_thr_given
+            END as snd_given
         from {self.source_tablename}
         LATERAL VIEW explode({self.get_column('children_cases')}) as children
         WHERE month='{self.month}'
         """)
+        return df1
 
     def aggregate_data(self, df):
         df.createOrReplaceTempView('pass1')
@@ -62,7 +66,8 @@ class ChildTHRAggregationHelper(BaseAggregationHelper):
             '{self.domain}' as domain,
             '{self.month}' as month,
             max(timeend) as last_timeend_processed,
-            sum(days_thr_given) as days_thr_given
+            sum(pse_attended) as total_pse_attended,
+            sum(snd_given) as total_snd_given
         FROM pass1
         GROUP BY child_case_id
         """)
