+++ dimagi_lake/aggregation/nutrition_project/table_mapping.py
@@ -1,0 +1,15 @@
+from dimagi_lake.aggregation.nutrition_project.agg_table_schema import (ChildCareMonthly,
+                                                                        FlwcLocation,
+                                                                        ServiceEnrollment,
+                                                                        ChildWeightHeightForm,
+                                                                        SupplementaryNutritionForm,
+                                                                        ChildTHRForm)
+
+TABLE_MAPPING = {
+    'location': FlwcLocation,
+    'child_care': ChildCareMonthly,
+    'service_enrollment': ServiceEnrollment,
+    'child_weight_height': ChildWeightHeightForm,
+    'supplementary_nutrition': SupplementaryNutritionForm,
+    'child_thr': ChildTHRForm
+}
