--- dimagi_lake/aggregation/table_mapping.py
@@ -1,15 +1,0 @@
-from dimagi_lake.aggregation.agg_table_schema import (ChildCareMonthly,
-                                                      FlwcLocation,
-                                                      ServiceEnrollment,
-                                                      ChildWeightHeightForm,
-                                                      SupplementaryNutritionForm,
-                                                      ChildTHRForm)
-
-TABLE_MAPPING = {
-    'location': FlwcLocation,
-    'child_care': ChildCareMonthly,
-    'service_enrollment': ServiceEnrollment,
-    'child_weight_height': ChildWeightHeightForm,
-    'supplementary_nutrition': SupplementaryNutritionForm,
-    'child_thr': ChildTHRForm
-}
